backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.4.2
prometheus-fastapi-instrumentator==6.1.0
requests==2.31.0
python-dotenv==1.0.0
opencv-python-headless==4.8.0.74
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.8